+++ e2e/tests/plugins.e2e.js
@@ -361,6 +361,7 @@ describe('plugins', () => {
       it('render upload modals', function() {
         embedTypes.forEach(embedType => {
           cy.openEmbedModal(STATIC_TOOLBAR_BUTTONS[embedType]);
+          cy.get(`[data-hook=verticalsImage]`).eq(3);
           cy.eyesCheckWindow(this.test.title);
           cy.get(`[data-hook*=settingPanelFooterCancel][tabindex!=-1]`).click();
         });